ci: add cl pipeline (#1)
* ci: first attempt to add build pipeline
* fix: implicit narrowing
* feat: allow to ignore decoder while deserializing
* ci: add libmusly to wheel when building binary distribution packages
* feat: add fallback to ffmpeg-cli for decoding when none decoder present
* ci: add first draft on tests
* doc: add draft documentation

+++ native/pymusly/pystream.h
@@ -262,7 +262,7 @@ class streambuf : public std::basic_streambuf<char>
       }
       farthest_pptr = (std::max)(farthest_pptr, pptr());
       off_type n_written = (off_type)(farthest_pptr - pbase());
-      pybind11::bytes chunk(pbase(), n_written);
+      pybind11::bytes chunk(pbase(), (Py_ssize_t)n_written);
       py_write(chunk);
       if (!traits_type::eq_int_type(c, traits_type::eof())) {
         char cs = traits_type::to_char_type(c);
